requirements.txt
flask==2.0.3
werkzeug==2.0.3
cohere==3.4.0
sentence-transformers==2.2.2
transformers==4.28.1
torch==2.2.2
faiss-cpu==1.7.3
pandas==1.5.3
numpy==1.24.1
requests==2.28.2
openpyxl==3.1.0